requirements.txt
agno
requests
beautifulsoup4
python-dotenv 